# blas/opencl cleanup