--- a/src/neural/network_blas.cc
+++ b/src/neural/network_blas.cc
@@ -77,87 +77,85 @@ class BlasComputation : public NetworkComputation {
     q_value_.emplace_back(std::tanh(winrate));
   }
 
-  void forward(std::vector<float>& input, std::vector<float>& output_pol,
+  void forward(std::vector<float>& input,
+               std::vector<float>& output_pol,
                std::vector<float>& output_val) {
+
+    ////////////////////////////////////////////////////////////////////////////
     // Input convolution
+
     constexpr int width = 8;
     constexpr int height = 8;
     constexpr int tiles = width * height / 4;
 
-    int NUM_VALUE_INPUT_PLANES = weights_.value.bn_means.size();
-    int NUM_POLICY_INPUT_PLANES = weights_.policy.bn_means.size();
-
-    static constexpr auto kWinogradAlpha = 4;
-    static constexpr auto kWinogradTile = kWinogradAlpha * kWinogradAlpha;
-
-    // Calculate output channels
-    const auto output_channels = weights_.input.biases.size();
-    // input_channels is the maximum number of input channels of any
-    // convolution.
+    // input_channels is the max number of input channels of any convolution.
     // Residual blocks are identical, but the first convolution might be bigger
     // when the network has very few filters
-    const auto input_channels = std::max(static_cast<size_t>(output_channels),
+    const auto input_channels = std::max(static_cast<size_t>(weights_.input.biases.size()),
                                          static_cast<size_t>(kInputPlanes));
-    auto conv_out = std::vector<float>(output_channels * width * height);
 
-    auto V = std::vector<float>(kWinogradTile * input_channels * tiles);
-    auto M = std::vector<float>(kWinogradTile * output_channels * tiles);
+    std::vector<float> conv_out(weights_.input.biases.size() * width * height);
 
-    std::vector<float> policy_data(NUM_POLICY_INPUT_PLANES * width * height);
-    std::vector<float> value_data(NUM_VALUE_INPUT_PLANES * width * height);
+    std::vector<float> V(kWinogradTile * input_channels * tiles);
+    std::vector<float> M(kWinogradTile * weights_.input.biases.size() * tiles);
 
-    Transforms::WinogradConvolve3(output_channels, input,
+    std::vector<float> policy_data(weights_.policy.bn_means.size() * width * height);
+    std::vector<float> value_data(weights_.value.bn_means.size() * width * height);
+
+    Transforms::WinogradConvolve3(weights_.input.biases.size(), input,
                                   weights_.input.weights, V, M, conv_out);
-    Transforms::Batchnorm<64>(output_channels, conv_out,
+    Transforms::Batchnorm<64>(weights_.input.biases.size(), conv_out,
                               weights_.input.bn_means.data(),
                               weights_.input.bn_stddivs.data());
 
+    ////////////////////////////////////////////////////////////////////////////
     // Residual tower
-    auto conv_in = std::vector<float>(output_channels * width * height);
-    auto res = std::vector<float>(output_channels * width * height);
+
+    std::vector<float> conv_in(weights_.residual[0].conv1.biases.size() * width * height);
+    std::vector<float> res(weights_.residual[0].conv1.biases.size() * width * height);
 
     for (auto& residual : weights_.residual) {
       auto& conv1 = residual.conv1;
-      auto output_channels = conv1.biases.size();
       std::swap(conv_out, conv_in);
       std::copy(begin(conv_in), end(conv_in), begin(res));
 
-      Transforms::WinogradConvolve3(output_channels, conv_in, conv1.weights, V,
+      Transforms::WinogradConvolve3(conv1.biases.size(), conv_in, conv1.weights, V,
                                     M, conv_out);
-      Transforms::Batchnorm<64>(output_channels, conv_out,
+      Transforms::Batchnorm<64>(conv1.biases.size(), conv_out,
                                 conv1.bn_means.data(), conv1.bn_stddivs.data());
 
       auto& conv2 = residual.conv2;
-      output_channels = conv2.biases.size();
       std::swap(conv_out, conv_in);
-      Transforms::WinogradConvolve3(output_channels, conv_in, conv2.weights, V,
+      Transforms::WinogradConvolve3(conv2.biases.size(), conv_in, conv2.weights, V,
                                     M, conv_out);
-      Transforms::Batchnorm<64>(output_channels, conv_out,
+      Transforms::Batchnorm<64>(conv2.biases.size(), conv_out,
                                 conv2.bn_means.data(), conv2.bn_stddivs.data(),
                                 res.data());
     }
 
-    Transforms::Convolve<1>(NUM_POLICY_INPUT_PLANES, conv_out,
+    ////////////////////////////////////////////////////////////////////////////
+    // Value/policy heads
+
+    Transforms::Convolve<1>(weights_.policy.bn_means.size(), conv_out,
                             weights_.policy.weights, weights_.policy.biases,
                             policy_data);
 
-    Transforms::Convolve<1>(NUM_VALUE_INPUT_PLANES, conv_out,
+    Transforms::Convolve<1>(weights_.value.bn_means.size(), conv_out,
                             weights_.value.weights, weights_.value.biases,
                             value_data);
 
-    Transforms::Batchnorm<width * height>(NUM_POLICY_INPUT_PLANES, policy_data,
+    Transforms::Batchnorm<width * height>(weights_.policy.bn_means.size(), policy_data,
                                           weights_.policy.bn_means.data(),
                                           weights_.policy.bn_stddivs.data());
 
-    Transforms::Batchnorm<width * height>(NUM_VALUE_INPUT_PLANES, value_data,
+    Transforms::Batchnorm<width * height>(weights_.value.bn_means.size(), value_data,
                                           weights_.value.bn_means.data(),
                                           weights_.value.bn_stddivs.data());
 
-    // NUM_POLICY_INPUT_PLANES*width*height x NUM_OUTPUT_POLICY
+
     Transforms::Innerproduct(policy_data, weights_.ip_pol_w, weights_.ip_pol_b,
                              output_pol);
 
-    // NUM_VALUE_INPUT_PLANES*width*height x NUM_VALUE_CHANNELS,
     Transforms::Innerproduct(value_data, weights_.ip1_val_w, weights_.ip1_val_b,
                              output_val, true);
   }
@@ -193,28 +191,28 @@ class BlasNetwork : public Network {
     bool verbose = options.GetOrDefault<bool>("verbose", true);
     int blas_cores = options.GetOrDefault<int>("blas_cores", 1);
 
-    const int inputChannels = kInputPlanes;
-    const int channels = weights.input.biases.size();
-    const size_t residual_blocks = weights.residual.size();
+    ////////////////////////////////////////////////////////////////////////////
+    // Input convolution
 
     weights_.input.weights = Transforms::WinogradTransformF(
-        weights_.input.weights, channels, inputChannels);
+        weights_.input.weights, weights.input.biases.size(), kInputPlanes);
 
     Transforms::OffsetBatchNormMeans(weights_.input.bn_means,
                                      weights_.input.biases);
 
     Transforms::InvertBatchNormStddev(weights_.input.bn_stddivs);
 
-    // residual blocks
-    for (size_t i = 0; i < residual_blocks; i++) {
-      auto& residual = weights_.residual[i];
+    ////////////////////////////////////////////////////////////////////////////
+    // Residual tower
+
+    for (auto& residual : weights_.residual) {
       auto& conv1 = residual.conv1;
       auto& conv2 = residual.conv2;
 
       conv1.weights =
-          Transforms::WinogradTransformF(conv1.weights, channels, channels);
+          Transforms::WinogradTransformF(conv1.weights, conv1.biases.size(), conv1.biases.size());
       conv2.weights =
-          Transforms::WinogradTransformF(conv2.weights, channels, channels);
+          Transforms::WinogradTransformF(conv2.weights, conv2.biases.size(), conv2.biases.size());
 
       Transforms::OffsetBatchNormMeans(conv1.bn_means, conv1.biases);
       Transforms::OffsetBatchNormMeans(conv2.bn_means, conv2.biases);
@@ -222,6 +220,9 @@ class BlasNetwork : public Network {
       Transforms::InvertBatchNormStddev(conv1.bn_stddivs);
       Transforms::InvertBatchNormStddev(conv2.bn_stddivs);
     }
+
+    ////////////////////////////////////////////////////////////////////////////
+    // Value/policy heads
 
     Transforms::OffsetBatchNormMeans(weights_.policy.bn_means,
                                      weights_.policy.biases);