1) a few renames to make things conceptually cleaner

2) remove unnecessary code/variables