@mooskagh 2018-06-05
Review status: 3 of 21 files reviewed at latest revision, 3 unresolved discussions.

---

*[src/neural/network_blas.cc, line 19 at r1](https://reviewable.io/reviews/leelachesszero/lc0/39#-LEGKvOlETUjB0UUjRh_:-LEGKvOlETUjB0UUjRha:bslnnax) ([raw file](https://github.com/leelachesszero/lc0/blob/906106725dc296625f611aaf516f3f3f0bf1073b/src/neural/network_blas.cc#L19)):*
> ```cpp
>  */
> 
> #include "neural/network.h"
> ```

Maybe move that file to neural/blas/ too? So that all files are together.
And then we can leave neural/ only only for single-file backends.

---

*[src/neural/network_opencl.cc, line 175 at r1](https://reviewable.io/reviews/leelachesszero/lc0/39#-LEGMCFu5LQAVsBKKfDX:-LEGMCFu5LQAVsBKKfDY:bls0a1v) ([raw file](https://github.com/leelachesszero/lc0/blob/906106725dc296625f611aaf516f3f3f0bf1073b/src/neural/network_opencl.cc#L175)):*
> ```cpp
>     // Residual tower
> 
>     for (auto& residual : weights.residual) {
> ```

`const auto&`
If I understand it correctly that `residual` isn't modified in the loop.

also in two lines below.

---

*[src/neural/network_opencl.cc, line 189 at r1](https://reviewable.io/reviews/leelachesszero/lc0/39#-LEGMZuf1i5FDi0BRB5q:-LEGMZuf1i5FDi0BRB5r:b-4s59z) ([raw file](https://github.com/leelachesszero/lc0/blob/906106725dc296625f611aaf516f3f3f0bf1073b/src/neural/network_opencl.cc#L189)):*
> ```cpp
>                                         m_ceil, m_ceil);
> 
>       std::vector<float> batchnorm_means_1 = conv1.bn_means;  // copy ctor
> ```

Could you add comment why it is necessary to have a copy?

---

*[src/neural/blas/transforms.h, line 65 at r1](https://reviewable.io/reviews/leelachesszero/lc0/39#-LEGFtdIBa_-mf5qUs0n-r1-65:-LEGLYlpC1_0AIOn8dt0:b-6zizey) ([raw file](https://github.com/leelachesszero/lc0/blob/906106725dc296625f611aaf516f3f3f0bf1073b/src/neural/blas/transforms.h#L65)):*
<details><summary><i>Previously, frpays (François Pays) wrote…</i></summary><blockquote>

I  kind of liked the const float* formulation, and at least, it was consistent with the 3rd argument.
</blockquote></details>

Could you add comment what eltwise really means?
It would also be nice to add comments to all public functions, but I'm not sure whether we decided to require code in blas/ and opencl/ to be styleguide-compliant or it's "semi-external" for now.

---

*[src/neural/blas/transforms.cc, line 370 at r1](https://reviewable.io/reviews/leelachesszero/lc0/39#-LEGNA6GEtJSZmLwJADs:-LEGNA6GEtJSZmLwJADt:b-8bwmo2) ([raw file](https://github.com/leelachesszero/lc0/blob/906106725dc296625f611aaf516f3f3f0bf1073b/src/neural/blas/transforms.cc#L370)):*
> ```cpp
> }
> 
> void BlasTransforms::OffsetBatchNormMeans(std::vector<float>& bn_means,
> ```

Actually, non-const reference parameters are disallowed by style guide.
If function changes argument, it should be pointer.

(I'm also fine with declaring neural/blas/ a styleguild-incompliant zone)

---

*[src/neural/opencl/cl2.hpp, line 1 at r1](https://reviewable.io/reviews/leelachesszero/lc0/39#-LEGQijR0OxuMjXsNwS5:-LEGQijR0OxuMjXsNwS6:blk7l8u) ([raw file](https://github.com/leelachesszero/lc0/blob/906106725dc296625f611aaf516f3f3f0bf1073b/src/neural/opencl/cl2.hpp#L1)):*
> ```cpp
> /*******************************************************************************
> ```

I thinks this should go from src/neural/opencl/cl2.hpp to third_party/CL/.
All code which we are not willing to touch ourselves should be there.

We can switch compiler to ignore all warnings in that code, and also add include there as -isystem rather than -I

Fine to do in separate PR, or I can do that myself later.

---


*Comments from [Reviewable](https://reviewable.io/reviews/leelachesszero/lc0/39)*
<!-- Sent from Reviewable.io -->