@mooskagh 2018-06-05
I didn't look into details of the conversation, but we should delete all unused code. If we ever need it, there is github history.

---

Review status: all files reviewed at latest revision, 8 unresolved discussions.

---



*Comments from [Reviewable](https://reviewable.io/reviews/leelachesszero/lc0/39#-:-LEGVD7Y11pyQ9tYtSMT:b-32ejqs)*
<!-- Sent from Reviewable.io -->