@dubslow 2018-06-05
Most of those comments should have been directed at the original porting PR. They can be addressed of course, but this PR isn't the appropriate place to discuss them.

The "removing unused but possibly useful-in-future code" discussion is the one where I'm most interested in your opinion. Obviously there's strong feelings on both sides of that particular discussion.