@killerducky 2018-06-05
Vote-wise it is 2:2 on the template thing. So crem needs to either decide to use himself as a tiebreaker, or get a 5th opinion and use that.

---

Review status: all files reviewed at latest revision, 8 unresolved discussions.

---



*Comments from [Reviewable](https://reviewable.io/reviews/leelachesszero/lc0/39#-:-LEGyRdfBBxHRfH-KClk:bmavhs4)*
<!-- Sent from Reviewable.io -->