@mooskagh 2018-06-06
Less code is bettet. Even if there is a large chance of using the code, like 70% (which is totally not the case here), I'd still say it's better to remove it.

In this particular case I'd remove it and left a comment in the code with a link to github review where original code could be found.

---

Review status: all files reviewed at latest revision, 9 unresolved discussions.

---



*Comments from [Reviewable](https://reviewable.io/reviews/leelachesszero/lc0/39#-:-LELq3ph-BlWCNQmKm4u:b-2t9oqz)*
<!-- Sent from Reviewable.io -->