@dubslow 2018-06-06
I'll make a revised version without the second commit (leaving `Transforms` name as is) and with the final comment leaving behind a lengthy comment and pointer to removed code.