@dubslow 2018-06-09
@frpays @killerducky I've left a comment block explaining the dead code situation, including a pointer to the commit that removed the code in question so that we may find it back as necessary. Please re-review.