@frpays 2018-06-24
I finished the new benchmarks

https://docs.google.com/spreadsheets/d/1eU2gb67V69_7yXFzIaF2WI4djF1pXGAih-rwhrstmg0


Platform | Threads | master | PR87 (512 batch size) | Speed-up
-- | -- | -- | -- | --
MacOS/vecLib | 1 | 43 | 114 | 2.65
MacOS/vecLib | 2 | 58 | 172 | 2.97
MacOS/vecLib | 4 | 89 | 222 | 2.49
MacOS/vecLib | 6 | 100 | 240 | 2.40
MacOS/vecLib | 8 | 98 | 239 | 2.44
MacOS/vecLib | 10 | 110 | 239 | 2.17
Linux/MKL 1-core | 1 | 24 | 38 | 1.58
Linux/MKL 1-core | 2 | 47 | 74 | 1.57
Linux/MKL 1-core | 4 | 89 | 134 | 1.51
Linux/MKL 1-core | 6 | 127 | 192 | 1.51
Linux/MKL 1-core | 8 | 161 | 245 | 1.52
Linux/MKL 1-core | 10 | 192 | 285 | 1.48
Linux/MKL 2-core | 1 | 33 | 60 | 1.82
Linux/MKL 2-core | 2 | 72 | 115 | 1.60
Linux/MKL 2-core | 4 | 132 | 213 | 1.61
Linux/MKL 2-core | 6 | 160 | 275 | 1.72
Linux/MKL 2-core | 8 | 98 | 250 | 2.55
Linux/MKL 2-core | 10 | 52 | 250 | 4.81
Linux/OpenBLAS 1-core | 1 | 24 | 39 | 1.63
Linux/OpenBLAS 1-core | 2 | 48 | 76 | 1.58
Linux/OpenBLAS 1-core | 4 | 89 | 140 | 1.57
Linux/OpenBLAS 1-core | 6 | 127 | 202 | 1.59
Linux/OpenBLAS 1-core | 8 | 160 | 257 | 1.61
Linux/OpenBLAS 1-core | 10 | 199 | 289 | 1.45
Linux/OpenBLAS 2-core | 1 | 29 | 61 | 2.10
Linux/OpenBLAS 2-core | 2 | 59 | 114 | 1.93
Linux/OpenBLAS 2-core | 4 | 102 | 207 | 2.03
Linux/OpenBLAS 2-core | 6 | 111 | 276 | 2.49
Linux/OpenBLAS 2-core | 8 | 100 | 256 | 2.56
Linux/OpenBLAS 2-core | 10 | 97 | 231 | 2.38